% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{find.Fidelity.linear}
\alias{find.Fidelity.linear}
\title{Title}
\usage{
find.Fidelity.linear(time.points, Fid.End, Fid.T1)
}
\arguments{
\item{Fid.T1}{}
}
\value{

}
\description{
calculates for given start Fidelity and end Fidelity the within
Fidelity for a given number of timepoints using a linear function
}
