% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{find.Fidelity.log}
\alias{find.Fidelity.log}
\title{Title}
\usage{
find.Fidelity.log(time.points, Fid.End, Fid.T1, par.slope = 1)
}
\arguments{
\item{par.slope}{}
}
\value{

}
\description{
Ccalculates for given start Fidelity and end Fidelity the within
Fidelity for a given number of timepoints using a logarithmus function,
where with the par.slope Stretching and compression of the function can be
influenced par.slope has to be greater than 0, near zero then the values are
almost constant, the greater then the values are going against the linear line.
}
