% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{calc.X.values}
\alias{calc.X.values}
\title{Approximate within points}
\usage{
calc.X.values(x.B, x.E, time.points)
}
\arguments{
\item{x.B}{starting point (begin)}

\item{x.E}{end point}

\item{time.points}{number of points on total}
}
\value{
all exact points
}
\description{
Calculation of all points in between a given start point and
end point for a given total number of points
}
