% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{calc.X.values}
\alias{calc.X.values}
\title{Title}
\usage{
calc.X.values(x.B, x.E, time.points)
}
\arguments{
\item{time.points}{}
}
\value{

}
\description{
Calculates between a given start point and and end point the points in between for a given total number of points
}
