% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{find.Fidelity.linear}
\alias{find.Fidelity.linear}
\title{Find approximate fidelity with a linear function}
\usage{
find.Fidelity.linear(time.points, Fid.End, Fid.T1)
}
\arguments{
\item{time.points}{number of time points}

\item{Fid.End}{Fidelity at the end (final time point)}

\item{Fid.T1}{Fidelity at the begin (starting time point)}
}
\value{
a matrix: first column with time points, second column corresponding Fidelity for the time points
}
\description{
calculates for given start Fidelity and end Fidelity the within
Fidelity for a given number of time points using a linear function
}
