% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulation}
\alias{simulation}
\title{Title}
\usage{
simulation(
  anzSim,
  type,
  sigma.1,
  sigma.2 = NULL,
  sigma.3,
  K,
  J,
  I,
  mu.0,
  theta,
  betas,
  X,
  X.A,
  B.cond = "0",
  C.cond = 0,
  D.cond = 0
)
}
\arguments{
\item{D.cond}{}
}
\value{

}
\description{
Title
}
